refactor: rename thank you page components
