+++ support-frontend/assets/pages/[countryGroupId]/components/thankYouComponent.tsx
